server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=14.0",
    "asyncpg>=0.30.0",
    "pyjwt>=2.9",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
]
